Moved dummylogger to Plumbing and some other requested changes.
