+++ Common.Logging.NLog45/Common.Logging.NLog45.csproj
@@ -16,7 +16,7 @@
 
   <ItemGroup>
     <PackageReference Include="Common.Logging" Version="3.4.1" />
-    <PackageReference Include="NLog" Version="4.7.7" />
+    <PackageReference Include="NLog" Version="4.7.10" />
   </ItemGroup>
 
 </Project>
