+++ Common.Logging.Scopes.Tests/AbstractScopesTests.cs
@@ -1,15 +1,7 @@
-﻿using System;
-using System.Linq;
-using System.Threading;
+﻿using System.Threading;
 using System.Threading.Tasks;
-using System.Collections.Generic;
-using System.Collections.Concurrent;
-using System.Runtime.Remoting.Messaging;
 
 using NUnit.Framework;
-
-using Common.Logging.Factory;
-using System.Collections.ObjectModel;
 
 namespace Common.Logging.Scopes
 {
@@ -30,441 +22,6 @@ namespace Common.Logging.Scopes
 		public AbstractScopesTests(bool useLogicalContexts)
 		{
 			_useLogicalContexts = useLogicalContexts;
-		}
-
-		#endregion
-
-		#region Inner Classes
-
-		private class DummyLog : AbstractLogger
-		{
-			#region AbstractLogger members..
-
-			public override bool IsTraceEnabled => false;
-
-			public override bool IsDebugEnabled => false;
-
-			public override bool IsInfoEnabled => false;
-
-			public override bool IsWarnEnabled => false;
-
-			public override bool IsErrorEnabled => false;
-
-			public override bool IsFatalEnabled => false;
-
-			protected override void WriteInternal(LogLevel level, object message, Exception exception)
-			{
-				// Do nothing..
-			}
-
-			#endregion
-
-			#region Fields & Properties
-
-			private readonly bool _useLogicalContexts;
-
-			private readonly IVariablesContext _threadVariablesContext = new VariablesContext();
-			private readonly IVariablesContext _logicalThreadVariablesContext = new LogicalVariablesContext();
-
-			private readonly INestedVariablesContext _nestedVariablesContext = new NestedVariablesContext();
-			private readonly INestedVariablesContext _logicalNestedVariablesContext = new LogicalNestedVariablesContext();
-
-			public override IVariablesContext ThreadVariablesContext => _useLogicalContexts ? _logicalThreadVariablesContext : _threadVariablesContext;
-			public override INestedVariablesContext NestedThreadVariablesContext => _useLogicalContexts ? _logicalNestedVariablesContext : _nestedVariablesContext;
-
-			#endregion
-
-			#region Inner Classes
-
-			private class VariablesContext : IVariablesContext
-			{
-				#region Fields
-
-				private readonly static ThreadLocal<IDictionary<string, object>> _context = new ThreadLocal<IDictionary<string, object>>(() => new Dictionary<string, object>());
-
-				#endregion
-
-				#region IVariablesContext members..
-
-				public object Get(string key) => _context.Value.ContainsKey(key) ? _context.Value[key] : null;
-
-				public bool Contains(string key) => _context.Value.ContainsKey(key);
-
-				public void Remove(string key) => _context.Value.Remove(key);
-
-				public void Clear() => _context.Value.Clear();
-
-				public void Set(string key, object value)
-				{
-					if (_context.Value.ContainsKey(key))
-						_context.Value[key] = value;
-					else
-						_context.Value.Add(key, value);
-				}
-
-				#endregion
-			}
-
-			private class NestedVariablesContext : INestedVariablesContext
-			{
-				#region Fields
-
-				private readonly static ThreadLocal<Stack<string>> _context = new ThreadLocal<Stack<string>>(() => new Stack<string>());
-
-
-				#endregion
-
-				#region INestedVariablesContext members..
-
-				public bool HasItems => _context.Value.Count > 0;
-
-				public string Peek() => _context.Value.Peek();
-
-				public void Clear() => _context.Value.Clear();
-
-				public string Pop() => _context.Value.Pop();
-
-				public IDisposable Push(string text)
-				{
-					_context.Value.Push(text);
-
-					return null; //< XXX: We don't need? To do here 'something' disposable to our tests btw..
-				}
-
-				#endregion
-			}
-
-			private class LogicalVariablesContext : IVariablesContext
-			{
-				#region Inner Classes
-
-				// XXX: Originally taken from MappedDiagnosticsLogicalContext.cs (NLog).
-				private static class Context
-				{
-					#region Inner Types
-
-					private sealed class ItemRemover : IDisposable
-					{
-						#region Fields
-
-						private readonly string _item;
-						private readonly bool _wasEmpty;
-
-						private int _disposed; //< Bool as int to allow the use of Interlocked.Exchange..
-
-						#endregion
-
-						#region .ctors
-
-						public ItemRemover(string item, bool wasEmpty)
-						{
-							_item = item;
-							_wasEmpty = wasEmpty;
-						}
-
-						#endregion
-
-						private bool RemoveScopeWillClearContext()
-						{
-							var immutableDict = GetLogicalThreadDictionary(false);
-
-							if (immutableDict.Count == 1 && immutableDict.ContainsKey(_item))
-								return true;
-
-							return false;
-						}
-
-						public void Dispose()
-						{
-							if (Interlocked.Exchange(ref _disposed, 1) == 0)
-							{
-								if (_wasEmpty && RemoveScopeWillClearContext())
-								{
-									Clear();
-									return;
-								}
-
-								var dictionary = GetLogicalThreadDictionary(true);
-								dictionary.Remove(_item);
-							}
-						}
-
-						public override string ToString()
-						{
-							return _item?.ToString() ?? base.ToString();
-						}
-					}
-
-					#endregion
-
-					#region Fields
-
-					private static readonly IDictionary<string, object> EmptyDefaultDictionary = new Dictionary<string, object>();
-
-					private const string LogicalThreadDictionaryKey = nameof(LogicalThreadDictionaryKey);
-
-					#endregion
-
-					#region Private methods
-
-					private static IDictionary<string, object> GetLogicalThreadDictionary(bool clone = false, int initialCapacity = 0)
-					{
-						var dictionary = GetThreadLocal();
-
-						if (dictionary == null)
-						{
-							if (!clone)
-								return EmptyDefaultDictionary;
-
-							dictionary = new Dictionary<string, object>(initialCapacity);
-
-							SetThreadLocal(dictionary);
-						}
-						else if (clone)
-						{
-							var newDictionary = new Dictionary<string, object>(dictionary.Count + initialCapacity);
-
-							foreach (var keyValue in dictionary)
-								newDictionary[keyValue.Key] = keyValue.Value;
-
-							SetThreadLocal(newDictionary);
-
-							return newDictionary;
-						}
-
-						return dictionary;
-					}
-
-					private static void SetThreadLocal(Dictionary<string, object> newValue)
-					{
-						if (newValue == null)
-							CallContext.FreeNamedDataSlot(LogicalThreadDictionaryKey);
-						else
-							CallContext.LogicalSetData(LogicalThreadDictionaryKey, newValue);
-					}
-
-					private static Dictionary<string, object> GetThreadLocal()
-					{
-						return CallContext.LogicalGetData(LogicalThreadDictionaryKey) as Dictionary<string, object>;
-					}
-
-					#endregion
-
-					public static void Set(string item, object value)
-					{
-						var logicalContext = GetLogicalThreadDictionary(true, 1);
-						logicalContext[item] = value;
-					}
-
-					public static object GetObject(string item)
-					{
-						GetLogicalThreadDictionary().TryGetValue(item, out var value);
-
-						return value;
-					}
-
-					public static bool Contains(string item)
-					{
-						return GetLogicalThreadDictionary().ContainsKey(item);
-					}
-
-					public static void Remove(string item)
-					{
-						GetLogicalThreadDictionary(true).Remove(item);
-					}
-
-					public static void Clear()
-					{
-						SetThreadLocal(null);
-					}
-				}
-
-				#endregion
-
-				#region IVariablesContext members..
-
-				public void Set(string key, object value) => Context.Set(key, value);
-
-				public bool Contains(string key) => Context.Contains(key);
-
-				public object Get(string key) => Context.GetObject(key);
-
-				public void Remove(string key) => Context.Remove(key);
-
-				public void Clear() => Context.Clear();
-
-				#endregion
-			}
-
-			private class LogicalNestedVariablesContext : INestedVariablesContext
-			{
-				#region Inner Classes
-
-				// XXX: Originally taken from NestedDiagnosticsLogicalContext.cs (NLog).
-				private static class Context
-				{
-					#region Inner Types
-
-					private interface INestedContext : IDisposable
-					{
-						INestedContext Parent { get; }
-
-						int FrameLevel { get; }
-						string Value { get; }
-					}
-
-					[Serializable]
-					private class NestedContext : INestedContext
-					{
-						#region Properties & Fields
-
-						public INestedContext Parent { get; }
-						public int FrameLevel { get; }
-						public string Value { get; }
-
-						private int _disposed;
-
-						#endregion
-
-						public static INestedContext CreateNestedContext(INestedContext parent, string value)
-						{
-							return new NestedContext(parent, value);
-						}
-
-						public NestedContext(INestedContext parent, string value)
-						{
-							Value = value;
-							Parent = parent;
-							FrameLevel = parent?.FrameLevel + 1 ?? 1;
-						}
-
-						public override string ToString()
-						{
-							return Value;
-						}
-
-						public void Dispose()
-						{
-							if (Interlocked.Exchange(ref _disposed, 1) != 1)
-								PopString();
-						}
-					}
-
-					#endregion
-
-					#region Fields
-
-					private const string NestedDiagnosticsContextKey = nameof(NestedDiagnosticsContextKey);
-
-					#endregion
-
-					#region Private methods
-
-					private static void SetThreadLocal(INestedContext newValue)
-					{
-						if (newValue == null)
-							CallContext.FreeNamedDataSlot(NestedDiagnosticsContextKey);
-						else
-							CallContext.LogicalSetData(NestedDiagnosticsContextKey, newValue);
-					}
-
-					private static INestedContext GetThreadLocal()
-					{
-						return CallContext.LogicalGetData(NestedDiagnosticsContextKey) as INestedContext;
-					}
-
-					#endregion
-
-					public static IDisposable Push(string value)
-					{
-						var parent = GetThreadLocal();
-						var current = NestedContext.CreateNestedContext(parent, value);
-
-						SetThreadLocal(current);
-
-						return current;
-					}
-
-					public static string[] GetAllMessages()
-					{
-						var currentContext = GetThreadLocal();
-
-						if (currentContext == null)
-							return new string[] { };
-
-						var index = 0;
-						var messages = new string[currentContext.FrameLevel];
-
-						while (currentContext != null)
-						{
-							messages[index++] = currentContext.Value;
-							currentContext = currentContext.Parent;
-						}
-
-						return messages;
-					}
-
-					public static string PopString()
-					{
-						var current = GetThreadLocal();
-
-						if (current != null)
-							SetThreadLocal(current.Parent);
-
-						return current?.Value;
-					}
-
-					public static string Peek()
-					{
-						return GetThreadLocal()?.Value;
-					}
-
-					public static string Pop()
-					{
-						return PopString() ?? string.Empty;
-					}
-
-					public static void Clear()
-					{
-						SetThreadLocal(null);
-					}
-				}
-
-				#endregion
-
-				#region INestedVariablesContext members..
-
-				public bool HasItems => Context.GetAllMessages().Length > 0;
-
-				public IDisposable Push(string text) => Context.Push(text);
-
-				public void Clear() => Context.Clear();
-
-				public string Peek() => Context.Peek();
-
-				public string Pop() => Context.Pop();
-
-				#endregion
-			}
-
-			#endregion
-
-			#region .ctors
-
-			public DummyLog(bool useLogicalContexts)
-			{
-				_useLogicalContexts = useLogicalContexts;
-			}
-
-			#endregion
-
-			// XXX: Expose our custom peek method for NestedVariablesContext.. that will facilitate test asserts.
-			protected internal string PeekFromNestedContext()
-			{
-				return _useLogicalContexts
-					? (NestedThreadVariablesContext as LogicalNestedVariablesContext).Peek()
-					: (NestedThreadVariablesContext as NestedVariablesContext).Peek();
-			}
 		}
 
 		#endregion
