+++ Common.Logging.Scopes.Tests/Common.Logging.Scopes.Tests.csproj
@@ -17,7 +17,8 @@
 
   <ItemGroup>
     <PackageReference Include="Common.Logging" Version="3.4.1" />
-    <PackageReference Include="NUnit" Version="3.13.1" />
+    <PackageReference Include="NLog" Version="4.7.10" />
+    <PackageReference Include="NUnit" Version="3.13.2" />
   </ItemGroup>
 
   <ItemGroup>
